New breed function, added normalize() and roll_dice()
A new breed function was written to replace the old one. The new one takes advantage of the new normalize() function, which is a step up from using topNPercent(). Also added was roll_dice(), which takes a probability (e.g. .50 for 50%) and returns True roughly that percentage of the time.

In addition to being used in breed(), mutate() was modified to use the roll_dice() function when determining whether a gene will mutate.

topNPercent() is commented out, as is the old breed function. They will be removed after more testing of the new functions.